% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_strata.R
\name{count_strata}
\alias{count_strata}
\title{Count number of strata in a StoX file}
\usage{
count_strata(path)
}
\arguments{
\item{path}{Path to the txt file}
}
\value{
The `number` of strata covered in the survey
}
\description{
A short description...
}
\examples{
path <- system.file("extdata/stox_winter_2019_2019/output/baseline/report/2_EstimateByPopulationCategory_Reports_Abundance.txt", package = "StoXreader")
count_strata(path)
}
