% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stox.R
\name{import_stox}
\alias{import_stox}
\title{Import and elongate individual StoX report files}
\usage{
import_stox(path, line_info, report_type)
}
\arguments{
\item{path}{Path to the txt file. Passed on from read_stox()}

\item{line_info}{A `data.frame` with three columns; n_skip, n_rows, and
stratum_n that provides information about table locations and their
associated stratum number.}

\item{report_type}{A character string to name value column after elongating.}
}
\value{
A `data.frame` with the number of lines to skip on import (n_skip),
the number of rows (n_rows), and stratum number (stratum_n).
}
\description{
import_stox() is a nested function in read_stox(). Its purpose is to import
the raw Stox files, then iterate over every table using the location
information given by line_info() and combine the tables into a single
data frame.
}
\examples{
path <- system.file(
  "extdata/stox_winter_2019/2_EstimateByPopulationCategory_Reports_Abundance.txt",
  package = "StoXreader")
line_info <- data.frame(
  n_skip = c(9, 51, 93, 135),
  n_rows = c(26, 26, 26, 26),
  stratum_n = c(1, 2, 3, 4))
report_type <- "Abundance"

import_stox(path, line_info, report_type)
}
