% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_startend.R
\name{locate_startend}
\alias{locate_startend}
\title{Locate start or end points of tables in StoX report file based on distinct
strings}
\usage{
locate_startend(vector, string, n_distance)
}
\arguments{
\item{vector}{Character vector with identifiable strings.}

\item{string}{String to locate in file.}

\item{n_distance}{Distance from string to start/end point.}
}
\value{
A `numeric vector object` with the line numbers of the start or
end points of tables in StoX report files.
}
\description{
locate_startend() is a nested function in find_all_strata() that takes a
character vector, string, and n_distance given by find_all_strata(), then
passes row numbers back to find_all_strata().
}
\examples{
vector <- c("LenGrp", NA, "__", "10-15", "LenGrp", "15-20", "___", "LenGrp")
string <- "LenGrp"
n_distance <- 1

locate_startend(vector, string, n_distance)
}
