% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stox_vector.R
\name{stox_vector}
\alias{stox_vector}
\title{Create character vector object from the first column of a StoX file}
\usage{
stox_vector(path)
}
\arguments{
\item{path}{Path to the txt file. Passed on from find_all_strata() or read_stox()}
}
\value{
A `character vector object`.
}
\description{
stox_vector() is a nested function in find_all_strata() and read_stox() that
takes a file path and converts the first column of the txt file into a
character vector so the number of lines can be compared.
}
\examples{
path <- system.file(
"stox_winter_2019/2_EstimateByPopulationCategory_Reports_Abundance.txt",
package = "StoXreader")

stox_vector(path)
}
