% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_age_s.R
\name{weight_age_s}
\alias{weight_age_s}
\title{Weight-age plot per stratum}
\usage{
weight_age_s(data, condition, facet = FALSE)
}
\arguments{
\item{data}{A data.frame with StoX data.}

\item{condition}{A character string with the name or number of a stratum.}

\item{facet}{Whether the plot should be wrapped into facets by year.}
}
\value{
A `ggplot` object with age on the x axis and mean weight on the y
axis. The points represent mean weight at specific ages.
}
\description{
A simple function that filters data to a single stratum and plots fish age
against mean weight.
}
\examples{
data <- cod
condition <- "1"

weight_age_s(data, condition)
}
