% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_age_y.R
\name{weight_age_y}
\alias{weight_age_y}
\title{Weight-age plot per year}
\usage{
weight_age_y(data, condition, facet = FALSE)
}
\arguments{
\item{data}{A data.frame with StoX data.}

\item{condition}{A character string or number with the year.}

\item{facet}{Whether the plot should be wrapped into facets by stratum.}
}
\value{
A `ggplot` object with age on the x axis and mean weight on the y
axis. The points represent mean weight at specific ages.
}
\description{
A simple function that filters data to a single year and plots fish age
against mean weight.
}
\examples{
data <- cod
condition <- "1999"

weight_age_y(data, condition)
}
