% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_strata.R
\name{find_all_strata}
\alias{find_all_strata}
\title{Find locations of all abundance tables per stratum}
\usage{
find_all_strata(path)
}
\arguments{
\item{path}{Path to the txt file}
}
\value{
A `data.frame`
}
\description{
A helper function designed to locate all tables in a StoX file if they are
separated by stratum. To extract the desired rows and columns, the location
to extract data from is needed.
}
\examples{
path <- system.file("extdata/stox_winter_2019_2019/output/baseline/report/2_EstimateByPopulationCategory_Reports_Abundance.txt", package = "StoXreader")
find_all_strata(path)
}
