% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stox.R
\name{read_stox}
\alias{read_stox}
\title{Read StoX report files}
\usage{
read_stox(a_path, b_path, w_path)
}
\arguments{
\item{a_path}{Path to abundance txt file.}

\item{b_path}{Path to biomass txt file.}

\item{w_path}{Path to mean weight txt file.}
}
\value{
A `data.frame`
}
\description{
Publicly available StoX files for Barents Sea Northeast Arctic cod bottom
trawl index in winter are divided into strata. The way these files are
formatted makes them require additional inputs to be converted into data
frames.

The three files in ~/output/baseline/report are reports on abundance,
biomass, and mean weight. read_stox() takes these three files, imports them,
then merges them into a single data frame.
}
\examples{
a_path <- system.file(
  "extdata/stox_winter_2019/2_EstimateByPopulationCategory_Reports_Abundance.txt",
  package = "StoXreader")
b_path <- system.file(
  "extdata/stox_winter_2019/2_EstimateByPopulationCategory_Reports_Biomass.txt",
  package = "StoXreader")
w_path <- system.file(
  "extdata/stox_winter_2019/2_EstimateByPopulationCategory_Reports_MeanWeight.txt",
  package = "StoXreader")

read_stox(a_path, b_path, w_path)
}
